requirements.txt
Django>=5.2.0
djangorestframework>=3.15.0
Pillow>=10.0.0
